Support for searching extra URL fields
There are Keepass clients like KeepassXC and Keepas2Android
that have added a way to assign multiple URLs to the same entry
by using some fields named KP2A_URL, KP2A_URL_1, KP2A_URL_2, etc.

This is a simple  change to include those fields in the results for
the search so that it's possible for them to show up.