Fix missing cached database when storage is low
The cached database is stored in a temporary folder which could be
cleared at any time
(https://pub.dev/documentation/path_provider/latest/path_provider/getTemporaryDirectory.html).
Unfortunately sometimes the cached database is no longer available when
the app is opened in offline mode.

This commit changed the directory for the cached database to the
Application Support Directory
(https://pub.dev/documentation/path_provider/latest/path_provider/getApplicationSupportDirectory.html).

The Application Documents Directory is not used because this directory
is already used for local databases and queried when opening local
databases.
Since the cached version of databases should not be selected manually
the Application Support Directory was chosen.